requirements.txt
asttokens==2.0.5
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
debugpy==1.6.2
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
idna==3.3
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.11.1
matplotlib-inline==0.1.3
natsort==8.1.0
nest-asyncio==1.5.5
numpy==1.23.1
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pyzmq==23.2.0
requests==2.28.1
six==1.16.0
stack-data==0.3.0
torch==1.12.0+cu113
torchaudio==0.12.0+cu113
torchsummary==1.5.1
torchvision==0.13.0+cu113
tornado==6.2
traitlets==5.3.0
typing-extensions==4.3.0
urllib3==1.26.10
wcwidth==0.2.5
